refactor: use fnmadd, move getx to test_utils
